requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.10.0
    # via
    #   httpx
    #   openai
    #   starlette
    #   watchfiles
asttokens==3.0.0
    # via stack-data
build==1.3.0
    # via pip-tools
certifi==2025.8.3
    # via
    #   httpcore
    #   httpx
    #   requests
    #   sentry-sdk
charset-normalizer==3.4.3
    # via requests
click==8.1.8
    # via
    #   pip-tools
    #   rich-toolkit
    #   typer
    #   uvicorn
decorator==5.2.1
    # via ipython
distro==1.9.0
    # via openai
dnspython==2.7.0
    # via email-validator
email-validator==2.2.0
    # via
    #   fastapi
    #   pydantic
exceptiongroup==1.3.0
    # via
    #   anyio
    #   ipython
executing==2.2.0
    # via stack-data
fastapi[all]==0.116.1
    # via -r requirements.in
fastapi-cli[standard]==0.0.8
    # via fastapi
fastapi-cloud-cli==0.1.5
    # via fastapi-cli
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via
    #   fastapi
    #   fastapi-cloud-cli
    #   openai
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
importlib-metadata==8.7.0
    # via build
ipython==8.18.1
    # via -r requirements.in
itsdangerous==2.2.0
    # via fastapi
jedi==0.19.2
    # via ipython
jinja2==3.1.6
    # via fastapi
jiter==0.10.0
    # via openai
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
matplotlib-inline==0.1.7
    # via ipython
mdurl==0.1.2
    # via markdown-it-py
openai==1.99.6
    # via -r requirements.in
orjson==3.11.1
    # via fastapi
packaging==25.0
    # via build
parso==0.8.4
    # via jedi
pexpect==4.9.0
    # via ipython
pip-tools==7.5.0
    # via -r requirements.in
prompt-toolkit==3.0.51
    # via ipython
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.3
    # via stack-data
pydantic[email]==2.11.7
    # via
    #   fastapi
    #   fastapi-cloud-cli
    #   openai
    #   pydantic-extra-types
    #   pydantic-settings
pydantic-core==2.33.2
    # via pydantic
pydantic-extra-types==2.10.5
    # via fastapi
pydantic-settings==2.10.1
    # via fastapi
pygments==2.19.2
    # via
    #   ipython
    #   rich
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
python-dotenv==1.1.1
    # via
    #   pydantic-settings
    #   uvicorn
python-multipart==0.0.20
    # via fastapi
pyyaml==6.0.2
    # via
    #   fastapi
    #   uvicorn
requests==2.32.4
    # via -r requirements.in
rich==14.1.0
    # via
    #   rich-toolkit
    #   typer
rich-toolkit==0.14.9
    # via
    #   fastapi-cli
    #   fastapi-cloud-cli
rignore==0.6.4
    # via fastapi-cloud-cli
sentry-sdk==2.34.1
    # via fastapi-cloud-cli
shellingham==1.5.4
    # via typer
sniffio==1.3.1
    # via
    #   anyio
    #   openai
stack-data==0.6.3
    # via ipython
starlette==0.47.2
    # via fastapi
tomli==2.2.1
    # via
    #   build
    #   pip-tools
tqdm==4.67.1
    # via openai
traitlets==5.14.3
    # via
    #   ipython
    #   matplotlib-inline
typer==0.16.0
    # via
    #   fastapi-cli
    #   fastapi-cloud-cli
typing-extensions==4.14.1
    # via
    #   anyio
    #   exceptiongroup
    #   fastapi
    #   ipython
    #   openai
    #   pydantic
    #   pydantic-core
    #   pydantic-extra-types
    #   rich-toolkit
    #   starlette
    #   typer
    #   typing-inspection
    #   uvicorn
typing-inspection==0.4.1
    # via
    #   pydantic
    #   pydantic-settings
ujson==5.10.0
    # via fastapi
urllib3==2.5.0
    # via
    #   requests
    #   sentry-sdk
uvicorn[standard]==0.35.0
    # via
    #   fastapi
    #   fastapi-cli
    #   fastapi-cloud-cli
uvloop==0.21.0
    # via uvicorn
watchfiles==1.1.0
    # via uvicorn
wcwidth==0.2.13
    # via prompt-toolkit
websockets==15.0.1
    # via uvicorn
wheel==0.45.1
    # via pip-tools
zipp==3.23.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
